move DetailsBox
