+++ frontend/src/components/common/index.js
@@ -1,3 +1,4 @@
+export * from './ColumnSelector'
 export * from './CourseDetails'
 export * from './ErrorBox'
 export * from './GoogleAnalytics'
